updated Interface Controller with Emojis
this is the updated interface controller with emoji buttons in the group, also i added the title outlet the vibe as well as all of the drag and drop updates i made to mine. Idk why but my new IC is much longer. might've been all of the drag and drops, guess ill say that for now